Remove literal kinds (#202)
Remove literal kinds.

+++ src/opsinputs/opsinputs_varobswriter_mod.F90
@@ -1479,14 +1479,14 @@ character(len=*), intent(in)                    :: JediVarName
 
 ! Local arguments:
 character(len=max_varname_with_channel_length) :: JediVarNamesWithChannels(max(size(Channels), 1))
-character(len=MAXVARLEN)        :: satidname
-real(kind=c_double)             :: VarValue(NumObs)
-real(kind=c_double)             :: MissingDouble
-integer(kind=4)                 :: SatIdValue(NumObs)
-integer(kind=4), allocatable    :: UniqueSatIds(:)
-integer                         :: ii, jj
-integer, parameter              :: maxpred = 31
-character(len=*), parameter     :: PredictorBaseName(1:maxpred) = (/ &
+character(len=MAXVARLEN)         :: satidname
+real(kind=c_double)              :: VarValue(NumObs)
+real(kind=c_double)              :: MissingDouble
+integer(kind=c_int)              :: SatIdValue(NumObs)
+integer(kind=c_int), allocatable :: UniqueSatIds(:)
+integer                          :: ii, jj
+integer, parameter               :: maxpred = 31
+character(len=*), parameter      :: PredictorBaseName(1:maxpred) = (/ &
               "constant                  ", &
               "thickness_850_300hPa      ", &
               "thickness_200_50hPa       ", &
@@ -1628,11 +1628,11 @@ end subroutine opsinputs_varobswriter_populatecxheader
 subroutine unique_values(input, output, positive)
 implicit none
 
-integer(kind=4), intent(in)               :: input(:)
-integer(kind=4), allocatable, intent(out) :: output(:)
-logical, optional, intent(in)             :: positive
+integer(kind=c_int), intent(in)               :: input(:)
+integer(kind=c_int), allocatable, intent(out) :: output(:)
+logical, optional, intent(in)                 :: positive
 
-integer(kind=4), allocatable :: unique(:)
+integer(kind=c_int), allocatable :: unique(:)
 integer :: i, j, k
 
 if (size(input) > 0) then
